Segmentation/model_maskrcnn/requirements.txt

#### prediction
pytorch
torchvision
pytorch_lightning
numpy
pandas 
Pillow
tqdm

##### training
# neptune-client
# efficientnet_pytorch

##### evaluation
# pycocotools
# ipyplot
